generify slab. use f32 elements instead of StrokePoint
