+++ Cargo.toml
@@ -29,6 +29,7 @@ human_bytes = "0.4.3"
 id_tree = "1.8.0"
 image = "0.24.6"
 log = "0.4.18"
+optional = "0.5.0"
 parking_lot = "0.12.1"
 raw-window-handle = "0.5.0"
 rayon = "1.7.0"
